adds FixEndpoint template class
