+++ src/libs/socket/FixClient.hpp
@@ -9,23 +9,6 @@
 
 #pragma once
 #include "Client.hpp"
-#include "fix/FixMessage.hpp"
-#include "fix/FixTag.hpp"
-#include <atomic>
-#include <string>
+#include "FixEndpoint.hpp"
 
-
-class FixClient : public Client
-{
-public:
-    FixClient() = default;
-
-    /* Thread-safe broadcast to all servers */
-    bool broadcast(FixMessage &message);
-
-protected:
-    std::string sendingTimeUTC() const;
-
-private:
-    std::atomic<long> _msgSeqNo{1};
-};
+using FixClient = FixEndpoint<Client>;
